boty-old/pyproject.toml
[tool.poetry]
name = "bots"
version = "0.1.0"
description = ""
authors = ["s24293 <s24293@pjwstk.edu.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
discord-py = "^2.3.2"
python-decouple = "^3.8"
transformers = "^4.39.2"
assemblyai = "^0.23.1"
torch = "^2.2.2"
openai = "^1.14.3"
python-telegram-bot = "13.7"
sentencepiece = "^0.2.0"
python-docx = "^1.1.0"
mysql-connector-python = "^8.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
